Added makefile for pushing to server, completed example compiler, moved to 64bit ints

+++ src/just-nums/src/main.rs
@@ -9,7 +9,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let program_text = fs::read_to_string(filename)?;
     // println!("Input = {}", program_text);
 
-    let program = i32::from_str(&program_text.trim())?;
+    let program = i64::from_str(&program_text.trim())?;
     println!("section .text\n\
 global our_code_starts_here\n\
 our_code_starts_here:\n\
